cool update
Implemented Depth and Breadth first search. In addition, I included the graph's methods to print the results as an Adjacency List and in matrix form. The structure of my current code is not ideal, too many things are hard coded. Will work on generalizing the process of reading the graph text file and executing the traversals.